Updated Device isolation icon for Linux
Device isolation is in preview for Linux according to the https://learn.microsoft.com/en-us/microsoft-365/security/defender-endpoint/respond-machine-alerts?view=o365-worldwide#isolate-devices-from-the-network and is available in console for Linux machine too. In the docs there is no - red cross for this feature, I propose switching it to yes - green check with a reference to preview features article.